Add lazy find and finds  #58 Add exists to lazy element #54
