+++ Framework/NugetBuild/packageSpecFlow.nuspec
@@ -13,6 +13,12 @@
     <iconUrl>http://magenicautomation.azurewebsites.net/content/favicon.ico</iconUrl>
     <summary>Magenic's automation quick start framework - SpecFlow extension. MAQS is a modular test automation framework that can be used as the base for your automation project or individual pieces can be used to enhance existing frameworks. Utilizing project templates, NuGet, and utility libraries we are able to have a functioning framework up and running in minutes, almost entirely removing on the initial time investment on implementing an automation solution.</summary>
     <releaseNotes>
+      5.7.0 - 2019/11/??????
+      Bug fixes and minor enhancements
+      - Have lazy element FindElement and FindElements return lazy element
+      - Add Exists and ExistsNow checks to lazy element
+      - Consolidate core lazy element and core lazy mobile element functionality
+
       5.6.0 - 2019/10/10
       Bug fixes and minor enhancements
       - Logging improvements
